test(router): move SGLang port regression into e2e module
